+++ client/src/pages/CreateEvent.jsx
@@ -17,15 +17,6 @@ export default function CreateEvent({ history }) {
 
   useEffect(() => {
     window.scrollTo(0, 0);
-    const onUnload = (e) => {
-      e.preventDefault();
-      window.scrollTo(0, 0);
-      e.returnValue = "이 페이지를 벗어나면 정성스럽게 작성한 글이 날아가요.";
-    };
-
-    window.addEventListener("beforeunload", onUnload);
-
-    return () => window.removeEventListener("beforeunload", onUnload);
   }, []);
 
   useEffect(() => {
