+++ client/src/style/theme.scss
@@ -95,6 +95,9 @@ body.dark {
         #time-picker-helper-text {
           color: #fff;
         }
+        .MuiInput-underline:before {
+          border-bottom: 1px solid rgba(255, 255, 255, 0.42);
+        }
         .eventform-title {
           color: $dark-font-primary-color;
         }
